+++ Example-Swift/JudoKitSwiftExample/ViewController.swift
@@ -94,10 +94,17 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     
     var isTransactingApplePayPreAuth = false
     
+    let judoKitSession = JudoKit(token: token, secret: secret)
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        JudoKit.showSecurityMessage = true
+        JudoKit.theme.acceptedCardNetworks = [Card.Configuration(.Visa, 16), Card.Configuration(.MasterCard, 16), Card.Configuration(.Maestro, 16), Card.Configuration(.AMEX, 15)]
+        
+        self.judoKitSession.sandboxed(true)
+        
+        
+        JudoKit.theme.showSecurityMessage = true
         
         self.tableView.backgroundColor = UIColor.clearColor()
         
@@ -157,7 +164,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     }
     
     @IBAction func AVSValueChanged(theSwitch: UISwitch) {
-        JudoKit.avsEnabled = theSwitch.on
+        JudoKit.theme.avsEnabled = theSwitch.on
     }
     
     // TODO: need to think of a way to add or remove certain card type acceptance as samples
@@ -207,7 +214,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     
     func paymentOperation() {
         guard let ref = Reference(consumerRef: "payment reference") else { return }
-        JudoKit.payment(judoID, amount: Amount(decimalNumber: 35, currency: currentCurrency), reference: ref, completion: { (response, error) -> () in
+        self.judoKitSession.payment(judoID, amount: Amount(decimalNumber: 35, currency: currentCurrency), reference: ref, completion: { (response, error) -> () in
             self.dismissViewControllerAnimated(true, completion: nil)
             if let error = error {
                 if error.code == .UserDidCancel {
@@ -237,7 +244,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     
     func preAuthOperation() {
         guard let ref = Reference(consumerRef: "payment reference") else { return }
-        JudoKit.preAuth(judoID, amount: Amount(decimalNumber: 2, currency: currentCurrency), reference: ref, completion: { (response, error) -> () in
+        self.judoKitSession.preAuth(judoID, amount: Amount(decimalNumber: 2, currency: currentCurrency), reference: ref, completion: { (response, error) -> () in
             self.dismissViewControllerAnimated(true, completion: nil)
             if let error = error {
                 if error.code == .UserDidCancel {
@@ -266,7 +273,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     
     func createCardTokenOperation() {
         guard let ref = Reference(consumerRef: "payment reference") else { return }
-        JudoKit.registerCard(judoID, amount: Amount(decimalNumber: 1, currency: currentCurrency), reference: ref, completion: { (response, error) -> () in
+        self.judoKitSession.registerCard(judoID, amount: Amount(decimalNumber: 1, currency: currentCurrency), reference: ref, completion: { (response, error) -> () in
             self.dismissViewControllerAnimated(true, completion: nil)
             if let error = error {
                 if error.code == .UserDidCancel {
@@ -291,7 +298,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     
     func repeatPaymentOperation() {
         if let cardDetails = self.cardDetails, let payToken = self.paymentToken, let ref = Reference(consumerRef: "payment reference") {
-            JudoKit.tokenPayment(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, cardDetails: cardDetails, paymentToken: payToken, completion: { (response, error) -> () in
+            self.judoKitSession.tokenPayment(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, cardDetails: cardDetails, paymentToken: payToken, completion: { (response, error) -> () in
                 self.dismissViewControllerAnimated(true, completion: nil)
                 if let error = error {
                     if error.code == .UserDidCancel {
@@ -325,7 +332,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
     
     func repeatPreAuthOperation() {
         if let cardDetails = self.cardDetails, let payToken = self.paymentToken, let ref = Reference(consumerRef: "payment reference") {
-            JudoKit.tokenPreAuth(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, cardDetails: cardDetails, paymentToken: payToken, completion: { (response, error) -> () in
+            self.judoKitSession.tokenPreAuth(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, cardDetails: cardDetails, paymentToken: payToken, completion: { (response, error) -> () in
                 self.dismissViewControllerAnimated(true, completion: nil)
                 if let error = error {
                     if error.code == .UserDidCancel {
@@ -436,9 +443,9 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
         guard let ref = Reference(consumerRef: "consumer Reference") else { return }
         
         if self.isTransactingApplePayPreAuth {
-            JudoKit.applePayPreAuth(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, payment: payment, completion: completionBlock)
+            self.judoKitSession.applePayPreAuth(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, payment: payment, completion: completionBlock)
         } else {
-            JudoKit.applePayPayment(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, payment: payment, completion: completionBlock)
+            self.judoKitSession.applePayPayment(judoID, amount: Amount(decimalNumber: 30, currency: currentCurrency), reference: ref, payment: payment, completion: completionBlock)
         }
     }